Added the common, small fixes

+++ averich.vladimir/I2/istriplet.cpp
@@ -7,7 +7,7 @@ size_t isTriplet(const int arr[], size_t size)
   {
     throw std::length_error("Error triplet");
   }
-  int countOfTriplet = 0;
+  size_t countOfTriplet = 0;
   for (size_t i = 2; i < size; i++)
   {
     if (arr[i] < arr[i - 1])
